sna_graph_chatbot/agent.py
import os
from dotenv import load_dotenv
from google.adk.agents import Agent
from google.adk.tools import AgentTool, FunctionTool
from google.adk.models.lite_llm import LiteLlm
from .callback import save_session_callback
from . import tools_graph as tg


sna_agent = Agent(
    name="SNA_Graph_Assistant",
    model=LiteLlm(model="ollama_chat/llama3.1:8b"),
    instruction=(
        "You are a conversational chatbot for Social Network Analysis.\n"
        "You MUST communicate ONLY in English.\n"
        "The user will ask questions ONLY in English.\n\n"

        "CRITICAL RULES:\n"
        "- Do NOT translate anything.\n"
        "- Do NOT use any language other than English.\n"
        "- Tool names and parameters are in English and MUST remain unchanged.\n\n"

        "ABSOLUTE RULE (VERY IMPORTANT):\n"
        "- After ANY tool call, you MUST ALWAYS send a final text response to the user.\n"
        "- Never stop after a tool call.\n"
        "- Never return only tool output.\n\n"

        "INTENT HANDLING:\n"
        "- If the user greets you (e.g. 'hello', 'hi'), reply politely in English something like this 'Hello ! How Can I help you?'  WITHOUT calling any tools.\n"
        "- If the user asks about graph statistics, call graph_overview().\n"
        "- If the user asks about shortest path between two nodes, call shortest_path(u=<int>, v=<int>).\n"
        "- If the user asks for neighbors of a node, call get_node_neighbors(u=<int>).\n" 
        "- If the user specifically asks for ALL neighbors, set show_all=True in get_node_neighbors.\n"
        "- If the user asks about communities, largest community, or modularity, call louvain_communities().\n"
        "- If the user asks for friend recommendations or to suggest friends for a node, "
        "- call recommend_friends(u=<int>, k=<int>).\n"
        "- Default k to 5 if the user doesn't specify a number.\n"
        "- If the user asks about articulation points / cut vertices, call articulation_points_top_k(k=<int>). "
        "If they do not specify k, use k=10.\n"
        "- If the user asks about bridges / cut edges, call bridges_top_k(k=<int>). "
        "If they do not specify k, use k=10.\n"
        "- If the user asks for a bridge/articulation summary, call bridge_summary().\n\n"

        "TOOL RESULT HANDLING:\n"
        "- Use ONLY the values returned by the tools.\n"
        "- NEVER invent numbers.\n"
        "- Explain the recommendations. Mention that they are based on the Adamic-Adar index, which looks at shared connections.\n"
        "- NEVER modify tool outputs.\n\n"

        "After receiving tool results, explain them clearly in English."
    ),
    tools=[
        FunctionTool(tg.graph_overview),
        FunctionTool(tg.ego_network),
        FunctionTool(tg.get_node_neighbors),
        FunctionTool(tg.recommend_friends),
        FunctionTool(tg.centralities_top_k),
        FunctionTool(tg.clustering_stats),
        FunctionTool(tg.k_core_summary),
        FunctionTool(tg.louvain_communities),
        FunctionTool(tg.degree_assortativity),
        FunctionTool(tg.articulation_points_top_k),
        FunctionTool(tg.bridges_top_k),
        FunctionTool(tg.bridge_summary),
        FunctionTool(tg.top_k_by_degree),
        FunctionTool(tg.shortest_path),
        FunctionTool(tg.component_summary),
        FunctionTool(tg.diameter_estimate),
    ],
    after_agent_callback=save_session_callback,
)

root_agent = sna_agent